New AbstractTableViewController defines core attributes and methods that TableViewControllers must implement
This class sets up the entity, formType, dataName and typeName that are
used by the TableViewControllers inside of the methods which are
defines as abstract.

The abstract methods defines in this class define the common actions
that the TableViewControllers provide when inheriting from a subclass
of this file. They are marked as abstract to allow the actions to be
overriden in cases that custom functionality is needed.