New TableViewController implements the methods defined in the Abstract..
The constructor of this class takes in values from the Child and uses
them in the methods that it implements. The child class provides the
TableViewController with the entity information for the entity that it
managed inside this class.

The actions here impement the defaults for managing an Entity with the
TableView, this includes new, edit, order up/down and delete. These by
default provide a good solid implementation which Controllers can
inherit from and manage the Entity as a TableViewControler.

However, certain of the methods can be overridden if there is a need
to, which provides a good base, but room for extension.

The swapRowOrder function calls the changeRowOrder function, which is
defines as part of a custom EntityRepository. There is a
BaseEntityRepository which provides this functionality, but required
that a custom EntityRepository is created for each Entity managed by a
TableViewContoller. This custom entity then needs to Inherit the
BaseEntityRepository and has access to the custom methods.