Made the ProjectHistoryController new and edit actions enable support for File uploading
Inside the actions once the form is validated, each File is persisted.
Then the File Uploader service handles the File Entity preparation
before persiting, and handles the uploading of the actual file itself,
aswell as creating thumbnail images if applicable.

In the edit action the findEntityFiles method created earlier is used
to find all Files belonging to the Entity. Thumbnail are then added to
the $thumbnails array, which are either a thumbnail image if they
exist, or the full image if there is no thumbnail. The Files Entities
are then added to the ProjectHistory Entity. This enables the Files to
be managed and thumbnails displayed next to the Files in the view.