Made several big changed to the FileUploader service. Check extended description for details.
The FileUploader service now automatically support for the File Entity.
This means that Files can be linked to Entities and the uploading of
the actual files from the File Entities is automatically peformed by
the service.

All that needs to be done is that Entities which want to support
integration with Files and the FileUpload service, is for the Entity to
extend \Corvus\AdminBundle\Entity\FileUpload.

The FileUpload Entity has a $files ArrayCollection which inherits to
the child class. It also implements the FileAccessorInterface and
FileUploadInterface, which provides methods to the child class which
act as Accessors for the $files array and for finding the uploads
directory.

To get this to work in controllers, $entity->getFiles() can be used to
get all the files. Then they need to be persisted to make the
FileUploader handle the rest. In an edit action, the
->findEntityImages() method of the Entities repository can be used,
then each File needs to ->add() so it is added to the Entity and
displayed in the edit view.