[0.1.0] Updated all Admin page to use Bootstrap
All of the Admin pages have been updated. They are now all styled based
on twitter bootstrap. The Home page of the Dashboard now has actual
information about the project, including a description, features, and
support links.

The general settings has been split, so the Analytics and Change
password sections are now in subpages of the General settings.

The WorkHistory, ProjectHistory, Navigation and Skills sections have
been updated, instead of the Table that was there before, it has been
replaced with a twitter boostrap version, which looks a lot better.

As a final note, FontAwesome has also been used for the Icons in places
like the Navigation and in the Table in the aforementioned sections.