[0.1.0] Various Form Type improvements.
Changes the Form Types to use setDefaultOptions instead of
getDefaultOptions. Removed the label_attr from most of the fields as
its no longer used.

Changes the property_path option to mapped, as it breaks Symfony
without as the syntax has changed.

Removed the “:” characters from the end of the Labels as it looks
better without them and consistent.