[0.1.0] Added Assetic Config, logos, Images ...
Added a new Assetic Config to manage the Font awesome integration
aswell as enable .less support.

Added a layout.less file for the backend, and updated the
design/layout.css files and removed unneeded styling.

Added Corvus logos to be used in the backend, small one in the Corvus
bar at the top, and the big logo in the Home page.

Added jquery.min and common Corvus js file to handle common JS
functionality.