[0.1.0] Added Licence and Readme, updated config
Added the Corvus licence and Readme files.

Corvus is now licences under the GPL v3.

Removed unneeded Bundles from the configuration files.

Also added filed to be ignored in the .gitignore file now that assetic
is generating files and they are being placed in new folders in the web
root.