[0.1.0] New methods PortfolioInfoRepository + Fix
Added new methods to the PortfolioInfoRepository class to better handle
the creation of Navigation items. It is now more flexible and can
handle creating Navigation items for the fronted, aswell as Navigation
and SubNavigation on the backend.

Also removed the lifecycleCallbacks in the GeneralSettings.orml.yml
file as it is no longer needed as the FileUploader is now a service.

Updated the config.yml file to inject the right dependencies to the
PortfolioInfoRepository.

The AppKernel is also updated to check if the application is being run
from the Command Line, if so it enters request scope. This stops the
PortfolioInfoRepository from breaking when using assetic to dump assets.

Also removed the change password validator as it has been replaced with
a change password page and was hacky anyway.