[0.1.0] Made humans.txt, changed Corvus to 0.1.0
I have added a humans.txt file to provide information about the Project
aswell as give credit/thanks to contributors/people who have helped the
project.

Just as a nice note here, I would like to thank James Williams for all
his help and continues support while I was at university, it was not
forgotten and was certainly appreciated.

Corvus has also been downgraded to 0.1.0.
The reason for this was that when I started the project I was
increasing the version number practically every time I made a commit.

In an effort to standardise the project will attempt to follow semantic
versioning, as such I feel it is appropriate to start the project at
0.1.0.

All future major releases will also be tagged appropriately.