[Service] Added a Controller preExecute Listener
The preExecute Listener takes in the PortfolioInfoRepository
getTemplateChoice.
It will automatically resolve the response to the correct sub folder in
the Resources/views folder. This is only done for the FrontendBundle.
Can be used in conjunction with the Template annotation, which will
automatically attempt to render a response, then the listener will
correct it to the correct template.