[Controller] Using Template annotation and Resolving id to Entity
Using the Template annotation to automatically resolve an array
returned from a controller to the correct template response.

Also changed the ProjectHistory and WorkHistory controller actions to
resolve to an Entity using the Id automatically. This is a magic part
of symfony which just works.
It will lookup the Entity by the Id and resolve it to the variable you
specify using the TypeHinting.