[Core] P. Extension loads routes from Frontend and Active Nav works with Prod/Dev
The PortfolioInfoRepository extension was moved to the CoreBundle from
the FrontendBundle. So now the loading of the route collection has been
changed to load from the FrontendBundle so it functions properly.

The detection of the current/active route was updated to allow it to
work with the .htaccess rewriting the URL to remove the app.php, and to
work when using app_dev.php